Revert fp16 sed patch on python/ — let prepare_env.sh do it on platform install
Commit 15abd6e72 manually applied the bf16->fp16 sed patch to
python/sglang/srt/layers/attention/minicpm_{backend,sparse_utils}.py to
unblock local sglang launch under --quantization gptq_marlin. That was
a debugging shortcut: prepare_env.sh has the same sed patch as its own
install-time step, so baking it into the source files made the
prepare_env.sh sed silently no-op (zero matches) on the v23 platform run.

quant_config_validator flagged this as CRITICAL: "sed will silently
no-op, backend stays bf16, fp16 Marlin GEMM produces garbage. SUSPECTED
CAUSE of v21 platform acc=0."

Reverting the python/ files back to commit 2c3a941c9's state restores
the 7+3 bf16 lines that prepare_env.sh's sed expects to match. Local
runs of local_eval.sh still apply the patch via the auto-sed-and-revert
trap added in a4cbcfdc6, so no local regression.

+++ python/sglang/srt/layers/attention/minicpm_backend.py
@@ -263,7 +263,7 @@ class MiniCPMSparseBackend(AttentionBackend):
             kernel_topk = tilelang.math.next_power_of_2(kernel_topk) // 2
         kernel_topk = max(8, kernel_topk)
         # FIXME: Read from model config
-        dtype_str = "float16"
+        dtype_str = "bfloat16"
         self.decode_fused_kernels = {}
         self.prefill_fused_kernels = {}
         bucketed_pooled_k_len = _bucket_size(pooled_k_len)
@@ -582,13 +582,13 @@ class MiniCPMSparseBackend(AttentionBackend):
                 metadata = self.forward_metadata
                 compressed_k = torch.full(
                     (forward_batch.batch_size * self.max_context_len // self.k1_kernel_stride, self.head_group_num, self.head_dim),
-                    dtype=torch.float16,
+                    dtype=torch.bfloat16,
                     device=self.device,
                     fill_value=float('-inf')
                 )
                 compressed_k2 = torch.full(
                     (forward_batch.batch_size * self.max_context_len // self.k2_kernel_stride, self.head_group_num, self.head_dim), 
-                    dtype=torch.float16, 
+                    dtype=torch.bfloat16, 
                     device=self.device,
                     fill_value=float('-inf')
                 )
@@ -748,7 +748,7 @@ class MiniCPMSparseBackend(AttentionBackend):
                     k2_token_nums=forward_batch.batch_size
                     * self.max_context_len
                     // self.k2_kernel_stride,
-                    dtype=torch.float16,
+                    dtype=torch.bfloat16,
                     device=self.device,
                     max_context_length=self.max_context_len,
                     split_stage1=self.split_stage1,
@@ -1341,7 +1341,7 @@ class MiniCPMSparseBackend(AttentionBackend):
                             self.head_group_num,
                             self.head_dim,
                         ),
-                        dtype=torch.float16,
+                        dtype=torch.bfloat16,
                         device=self.device,
                     ),
                     "compress_k2": torch.zeros(
@@ -1350,7 +1350,7 @@ class MiniCPMSparseBackend(AttentionBackend):
                             self.head_group_num,
                             self.head_dim,
                         ),
-                        dtype=torch.float16,
+                        dtype=torch.bfloat16,
                         device=self.device,
                     ),
                     "k1.cu_seqlens": torch.zeros(
